Basic motor without recipe for now

+++ src/main/java/org/antarcticgardens/newage/Configurations.java
@@ -4,7 +4,7 @@ public class Configurations {
 
     public static int MAX_RODS_IN_DIRECTION = 32;
 
-    public static float SU_TO_ENERGY = (float) 240 / 8192; // Making compatible with create: crafts and additions
+    public static float SU_TO_ENERGY = 240f / 8192f; // Making compatibility with create: crafts and additions
 
     public static int MAX_COILS = 8;
 }
